* master: (441 commits)
  docs: update contributing stuff
  docs: update contributing stuff
  feat: make AppDataForProjects non optional
  5.9.0
  fix: lint
  feat(autoRepair): make stray backup stuff translateable
  feat(autoRepair): re-enable stray backup check
  feat(autoRepair): add a little bit of logging
  feat(autoRepair): make restoring orphaned tasks work
  feat(autoRepair): make restore from archive work
  feat(autoRepair): make _removeMissingIdsFromLists work
  feat(autoRepair): make _removeDuplicatesFromArchive work
  feat(autoRepair): make fix duplicate tasks work
  refactor(autoRepair): make dataRepair non async
  test(autoRepair): prepare more tests
  test(autoRepair): make app data mock available
  test(autoRepair): outline tests
  feat(autoRepair): trigger for data import and data init if data is broken
  fix: another read-only error #538
  test: make new error cases work
  ...

# Conflicts:
#	electron/main-window.ts
#	electron/main.ts
#	package.json
#	src/main.ts

+++ src/app/features/note/dialog-add-note-reminder/dialog-add-note-reminder.component.ts
@@ -1,11 +1,11 @@
-import {ChangeDetectionStrategy, Component, Inject} from '@angular/core';
-import {MAT_DIALOG_DATA, MatDialogRef} from '@angular/material/dialog';
-import {Note} from '../note.model';
-import {NoteService} from '../note.service';
-import {ReminderCopy} from '../../reminder/reminder.model';
-import {ReminderService} from '../../reminder/reminder.service';
-import {T} from '../../../t.const';
-import {throttle} from 'helpful-decorators';
+import { ChangeDetectionStrategy, Component, Inject } from '@angular/core';
+import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
+import { Note } from '../note.model';
+import { NoteService } from '../note.service';
+import { ReminderCopy } from '../../reminder/reminder.model';
+import { ReminderService } from '../../reminder/reminder.service';
+import { T } from '../../../t.const';
+import { throttle } from 'helpful-decorators';
 
 @Component({
   selector: 'dialog-add-note-reminder',
@@ -14,11 +14,11 @@ import {throttle} from 'helpful-decorators';
   changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class DialogAddNoteReminderComponent {
-  T = T;
-  dateTime: number;
+  T: typeof T = T;
+  dateTime?: number;
   title: string;
   isEdit: boolean;
-  reminder: ReminderCopy;
+  reminder?: ReminderCopy | null;
   note: Note;
 
   constructor(
@@ -28,11 +28,11 @@ export class DialogAddNoteReminderComponent {
     @Inject(MAT_DIALOG_DATA) public data: { note: Note },
   ) {
     this.note = this.data.note;
-    this.reminder = this.note.reminderId && {
-      ...this._reminderService.getById(this.data.note.reminderId)
-    };
+    this.reminder = this.note.reminderId
+      ? this._reminderService.getById(this.data.note.reminderId as string)
+      : null;
     this.isEdit = !!(this.reminder && this.reminder.id);
-    if (this.isEdit) {
+    if (this.isEdit && this.reminder) {
       this.dateTime = this.reminder.remindAt;
       this.title = this.reminder.title;
     } else {
@@ -49,7 +49,7 @@ export class DialogAddNoteReminderComponent {
       return;
     }
 
-    if (this.isEdit) {
+    if (this.isEdit && this.reminder) {
       this._noteService.updateReminder(
         this.note.id,
         this.reminder.id,
@@ -70,6 +70,9 @@ export class DialogAddNoteReminderComponent {
   // NOTE: throttle is used as quick way to prevent multiple submits
   @throttle(2000, {leading: true, trailing: false})
   remove() {
+    if (!this.reminder) {
+      throw new Error('No reminder');
+    }
     this._noteService.removeReminder(this.note.id, this.reminder.id);
     this.close();
   }
