+++ src/app/features/tasks/dialog-add-task-reminder/dialog-add-task-reminder.component.ts
@@ -1,13 +1,12 @@
-import {ChangeDetectionStrategy, Component, Inject} from '@angular/core';
-import {MAT_DIALOG_DATA, MatDialogRef} from '@angular/material/dialog';
-import {TaskService} from '../task.service';
-import {ReminderCopy} from '../../reminder/reminder.model';
-import {ReminderService} from '../../reminder/reminder.service';
-import {SnackService} from '../../../core/snack/snack.service';
-import {T} from '../../../t.const';
-import {AddTaskReminderInterface} from './add-task-reminder-interface';
-import {WorkContextService} from '../../work-context/work-context.service';
-import {throttle} from 'helpful-decorators';
+import { ChangeDetectionStrategy, Component, Inject } from '@angular/core';
+import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
+import { TaskService } from '../task.service';
+import { ReminderCopy } from '../../reminder/reminder.model';
+import { ReminderService } from '../../reminder/reminder.service';
+import { T } from '../../../t.const';
+import { AddTaskReminderInterface } from './add-task-reminder-interface';
+import { throttle } from 'helpful-decorators';
+import { Task } from '../task.model';
 
 @Component({
   selector: 'dialog-add-task-reminder',
@@ -16,19 +15,19 @@ import {throttle} from 'helpful-decorators';
   changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class DialogAddTaskReminderComponent {
-  T = T;
-  task = this.data.task;
-  reminder: ReminderCopy = this.task.reminderId && this._reminderService.getById(this.task.reminderId);
+  T: typeof T = T;
+  task: Task = this.data.task;
+  reminder?: ReminderCopy = this.task.reminderId
+    ? this._reminderService.getById(this.task.reminderId) || undefined
+    : undefined;
   isEdit: boolean = !!(this.reminder && this.reminder.id);
 
-  dateTime: number = this.reminder && this.reminder.remindAt;
-  isShowMoveToBacklog: boolean = (!this.isEdit && !!this.task.projectId);
+  dateTime?: number = this.reminder && this.reminder.remindAt;
+  isShowMoveToBacklog: boolean = (!this.isEdit && !!this.task.projectId && this.task.parentId === null);
   isMoveToBacklog: boolean = (this.isShowMoveToBacklog);
 
   constructor(
     private _taskService: TaskService,
-    private _snackService: SnackService,
-    private _workContextService: WorkContextService,
     private _reminderService: ReminderService,
     private _matDialogRef: MatDialogRef<DialogAddTaskReminderComponent>,
     @Inject(MAT_DIALOG_DATA) public data: AddTaskReminderInterface,
@@ -44,7 +43,7 @@ export class DialogAddTaskReminderComponent {
       return;
     }
 
-    if (this.isEdit) {
+    if (this.isEdit && this.reminder) {
       this._taskService.updateReminder(
         this.task.id,
         this.reminder.id,
@@ -65,6 +64,10 @@ export class DialogAddTaskReminderComponent {
   // NOTE: throttle is used as quick way to prevent multiple submits
   @throttle(2000, {leading: true, trailing: false})
   remove() {
+    if (!this.reminder || !this.reminder.id) {
+      console.log(this.reminder, this.task);
+      throw new Error('No reminder or id');
+    }
     this._taskService.removeReminder(this.task.id, this.reminder.id);
     this.close();
   }
